"ads2_m6_a1_1: 3/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ assignments/m6/ADS-2Week-1/Solution.java
@@ -24,7 +24,7 @@ class PageRank {
 			for(int j = 0; j < g.V(); j++) {
 				for(int each: g.adj(j)) {
 					if(each == i) {
-						pr += list[j] / g.outdegree(j);
+						pr += (double)(list[j] / g.outdegree(j));
 					}
 				}
 			}
